Refactor: Move MCP protocol handling to McpCommunicationService
- Renamed CommandLineService to McpCommunicationService for clarity
- Moved all MCP protocol logic from McpProxy to McpCommunicationService
- Extracted tool schema definitions, response building, and error handling
- Simplified McpProxy to pure business logic orchestration (59 lines)
- Added health check integration for initialize response
- Updated interfaces to support new architecture
